[IMP] deactivate mkserver processes not needed in 6.1 anymore

+++ openerp-scripts/mkserver.sh
@@ -19,8 +19,8 @@
 #!/bin/bash
 
 if [[ ! -d $OPENERP_CCORP_DIR ]]; then
-	echo "openerp-ccorp-scripts not installed."
-	exit 1
+    echo "openerp-ccorp-scripts not installed."
+    exit 1
 fi
 
 #~ Libraries import
@@ -36,16 +36,16 @@ INSTALL_LOG_PATH=/var/log/openerp
 INSTALL_LOG_FILE=$INSTALL_LOG_PATH/install.log
 
 if [[ ! -f $INSTALL_LOG_FILE ]]; then
-	mkdir -p $INSTALL_LOG_PATH
-	touch $INSTALL_LOG_FILE
+    mkdir -p $INSTALL_LOG_PATH
+    touch $INSTALL_LOG_FILE
 fi
 
 function log {
-	echo "$(date): $1" >> $INSTALL_LOG_FILE
+    echo "$(date): $1" >> $INSTALL_LOG_FILE
 }
 function log_echo {
-	echo $1
-	log "$1"
+    echo $1
+    log "$1"
 }
 log ""
 
@@ -96,7 +96,7 @@ if [ $installed_count = 0 ]; then
 elif [ $installed_count = 1 ]; then
     branch=$installed_branch
 else
-	branch=""
+    branch=""
     while [[ ! $branch =~ ^5\.0$ ]] && [[ ! $branch =~ ^6\.0$ ]] && [[ ! $branch =~ ^6\.1$ ]] && [[ ! $branch =~ ^trunk$ ]]; do
         read -p "You have installed several versions, choose the version for this server (5.0 / 6.0 / _6.1_ / trunk)? " branch
         if [[ $branch == "" ]]; then
@@ -109,13 +109,13 @@ fi
 log_echo "This server will use $branch branch."
 
 if [[ $branch = "5.0" ]]; then
-	branch="5.0"
+    branch="5.0"
 elif [[ $branch = "6.0" ]]; then
-	branch="6.0"
+    branch="6.0"
 elif [[ $branch = "6.1" ]]; then
-	branch="6.1"
+    branch="6.1"
 else
-	branch="trunk"
+    branch="trunk"
 fi
 log_echo ""
 
@@ -123,10 +123,10 @@ log_echo ""
 
 name=""
 while [[ $name == "" ]]; do
-	read -p "Enter the OpenERP server name: " name
-	if [[ $name == "" ]]; then
-		log_echo "The name cannot be blank."
-	fi
+    read -p "Enter the OpenERP server name: " name
+    if [[ $name == "" ]]; then
+        log_echo "The name cannot be blank."
+    fi
 done
 
 #~ Set the openerp port
@@ -134,15 +134,15 @@ port=""
 test="^[0-9]{2}$"
 mkdir -p /etc/openerp/ports
 while [[ $port == "" ]]; do
-	read -p "Enter the OpenERP server port (2 digits only): " port
-	if [[ ! $port =~ $test ]]; then
-		log_echo "The port has to contain exactly 2 digits."
-		port=""
-	elif [[ $(ls -1 /etc/openerp/ports/${port}_* | wc -l) != 0 ]]; then
-		tmp=$(ls /etc/openerp/ports/${port}_* | xargs -n1 basename)
-		log_echo "The port $port is already in use by the server $(tmp:3) (/etc/openerp/ports)."
-		port=""
-	fi
+    read -p "Enter the OpenERP server port (2 digits only): " port
+    if [[ ! $port =~ $test ]]; then
+        log_echo "The port has to contain exactly 2 digits."
+        port=""
+    elif [[ $(ls -1 /etc/openerp/ports/${port}_* | wc -l) != 0 ]]; then
+        tmp=$(ls /etc/openerp/ports/${port}_* | xargs -n1 basename)
+        log_echo "The port $port is already in use by the server $(tmp:3) (/etc/openerp/ports)."
+        port=""
+    fi
 done
 touch /etc/openerp/ports/${port}_${name}
 log_echo "Selected port is: $port"
@@ -168,20 +168,20 @@ done
 #~ Set the openerp admin password
 admin_passwd=""
 while [[ $admin_passwd == "" ]]; do
-	read -p "Enter the OpenERP administrator password: " admin_passwd
-	if [[ $admin_passwd == "" ]]; then
-		log_echo "The password cannot be empty."
-	else
-		read -p "Enter the OpenERP administrator password again: " admin_passwd2
-		log_echo ""
-		if [[ $admin_passwd == $admin_passwd2 ]]; then
-			log_echo "OpenERP administrator password set."
-		else
-			admin_passwd=""
-			log_echo "Passwords don't match."
-		fi
-	fi
-	log_echo ""
+    read -p "Enter the OpenERP administrator password: " admin_passwd
+    if [[ $admin_passwd == "" ]]; then
+        log_echo "The password cannot be empty."
+    else
+        read -p "Enter the OpenERP administrator password again: " admin_passwd2
+        log_echo ""
+        if [[ $admin_passwd == $admin_passwd2 ]]; then
+            log_echo "OpenERP administrator password set."
+        else
+            admin_passwd=""
+            log_echo "Passwords don't match."
+        fi
+    fi
+    log_echo ""
 done
 
 # Add openerp postgres user
@@ -209,9 +209,9 @@ sed -i "s#\\[NAME\\]#$name#g" /etc/init.d/openerp-server-$name >> $INSTALL_LOG_F
 sed -i "s#\\[USER\\]#openerp_$name#g" /etc/init.d/openerp-server-$name >> $INSTALL_LOG_FILE
 #~ Start server on boot
 if [[ $start_boot =~ ^[Yy]$ ]]; then
-	log_echo "Creating server rc rules..."
-	update-rc.d openerp-server-$name defaults >> $INSTALL_LOG_FILE
-	log_echo ""
+    log_echo "Creating server rc rules..."
+    update-rc.d openerp-server-$name defaults >> $INSTALL_LOG_FILE
+    log_echo ""
 fi
 
 log_echo "Creating openerp-server bin script..."
@@ -233,17 +233,20 @@ mkserver_addons_config
 sed -i "s#\\[ADDONS\\]#$addons_config#g" /etc/openerp/$branch/server/$name.conf >> $INSTALL_LOG_FILE
 
 
-log_echo "Creating openerp-server ssl files..."
-cp -a /etc/openerp/ssl/server.cnf-skeleton /etc/openerp/ssl/servers/$name.cnf >> $INSTALL_LOG_FILE
-sed -i "s#\\[NAME\\]#$name#g" /etc/openerp/ssl/servers/$name.cnf >> $INSTALL_LOG_FILE
-cd /etc/openerp/ssl/servers
-openssl req -newkey rsa:1024 -keyout tempkey.pem -keyform PEM -out tempreq.pem -outform PEM -config $name.cnf -passout pass:$admin_passwd
-openssl rsa -passin pass:$admin_passwd < tempkey.pem > server_key.pem
-echo 20$port > /etc/openerp/ssl/serial
-openssl ca -batch -in tempreq.pem -out server_crt.pem -config ../ca.cnf -passin pass:$admin_passwd
-rm -f tempkey.pem && rm -f tempreq.pem
-mv server_crt.pem ${name}_crt.pem
-mv server_key.pem ${name}_key.pem
+# For OpenERP 6.1+ the XML-RPC secure is not used any more, so don't initiate the certificates
+if [[ "$branch" =~ ^5\.0|6\.0$ ]]; then
+    log_echo "Creating openerp-server ssl files..."
+    cp -a /etc/openerp/ssl/server.cnf-skeleton /etc/openerp/ssl/servers/$name.cnf >> $INSTALL_LOG_FILE
+    sed -i "s#\\[NAME\\]#$name#g" /etc/openerp/ssl/servers/$name.cnf >> $INSTALL_LOG_FILE
+    cd /etc/openerp/ssl/servers
+    openssl req -newkey rsa:1024 -keyout tempkey.pem -keyform PEM -out tempreq.pem -outform PEM -config $name.cnf -passout pass:$admin_passwd
+    openssl rsa -passin pass:$admin_passwd < tempkey.pem > server_key.pem
+    echo 20$port > /etc/openerp/ssl/serial
+    openssl ca -batch -in tempreq.pem -out server_crt.pem -config ../ca.cnf -passin pass:$admin_passwd
+    rm -f tempkey.pem && rm -f tempreq.pem
+    mv server_crt.pem ${name}_crt.pem
+    mv server_key.pem ${name}_key.pem
+fi
 
 log_echo "Creating openerp-server log files..."
 mkdir -p /var/log/openerp/$name >> $INSTALL_LOG_FILE
@@ -253,31 +256,36 @@ log_echo "Creating openerp-web bin script..."
 cp -a /etc/openerp/$branch/web-client/bin-skeleton /usr/local/bin/openerp-web-$name >> $INSTALL_LOG_FILE
 sed -i "s#\\[NAME\\]#$name#g" /usr/local/bin/openerp-web-$name >> $INSTALL_LOG_FILE
 
-log_echo "Creating openerp-web init script..."
-cp -a /etc/openerp/$branch/web-client/init-skeleton /etc/init.d/openerp-web-$name >> $INSTALL_LOG_FILE
-sed -i "s#\\[NAME\\]#$name#g" /etc/init.d/openerp-web-$name >> $INSTALL_LOG_FILE
-sed -i "s#\\[USER\\]#openerp_$name#g" /etc/init.d/openerp-web-$name >> $INSTALL_LOG_FILE
-#~ Start web client on boot
-if [[ $start_boot =~ ^[Yy]$ ]]; then
-	log_echo "Creating web-client rc rules..."
-	update-rc.d openerp-web-$name defaults >> $INSTALL_LOG_FILE
-	log_echo ""
+# For OpenERP 6.1+ the web server is embedded, so no need to initiate or configure the process
+if [[ "$branch" =~ ^5\.0|6\.0$ ]]; then
+    log_echo "Creating openerp-web init script..."
+    cp -a /etc/openerp/$branch/web-client/init-skeleton /etc/init.d/openerp-web-$name >> $INSTALL_LOG_FILE
+    sed -i "s#\\[NAME\\]#$name#g" /etc/init.d/openerp-web-$name >> $INSTALL_LOG_FILE
+    sed -i "s#\\[USER\\]#openerp_$name#g" /etc/init.d/openerp-web-$name >> $INSTALL_LOG_FILE
+    #~ Start web client on boot
+    if [[ $start_boot =~ ^[Yy]$ ]]; then
+        log_echo "Creating web-client rc rules..."
+        update-rc.d openerp-web-$name defaults >> $INSTALL_LOG_FILE
+        log_echo ""
+    fi
+
+    log_echo "Creating openerp-web configuration file..."
+    cp -a /etc/openerp/$branch/web-client/web-client.conf-$branch-skeleton /etc/openerp/$branch/web-client/$name.conf
+    sed -i "s#\\[NAME\\]#$name#g" /etc/openerp/$branch/web-client/$name.conf
+    sed -i "s#\\[PORT\\]#23$port#g" /etc/openerp/$branch/web-client/$name.conf
+    sed -i "s#\\[SERVER_PORT\\]#21$port#g" /etc/openerp/$branch/web-client/$name.conf
+    if [[ $server_type == "production" ]]; then
+        sed -i "s/#\?[[:space:]]*\(dbbutton\.visible.*\)/dbbutton.visible = False/g" /etc/openerp/$branch/web-client/$name.conf
+    else
+        sed -i "s/#\?[[:space:]]*\(dbbutton\.visible.*\)/dbbutton.visible = True/g" /etc/openerp/$branch/web-client/$name.conf
+    fi
+
+    log_echo "Creating openerp-web log files..."
+    touch /var/log/openerp/$name/web-client-access.log
+    touch /var/log/openerp/$name/web-client-error.log
 fi
 
-log_echo "Creating openerp-web configuration file..."
-cp -a /etc/openerp/$branch/web-client/web-client.conf-$branch-skeleton /etc/openerp/$branch/web-client/$name.conf
-sed -i "s#\\[NAME\\]#$name#g" /etc/openerp/$branch/web-client/$name.conf
-sed -i "s#\\[PORT\\]#23$port#g" /etc/openerp/$branch/web-client/$name.conf
-sed -i "s#\\[SERVER_PORT\\]#21$port#g" /etc/openerp/$branch/web-client/$name.conf
-if [[ $server_type == "production" ]]; then
-	sed -i "s/#\?[[:space:]]*\(dbbutton\.visible.*\)/dbbutton.visible = False/g" /etc/openerp/$branch/web-client/$name.conf
-else
-	sed -i "s/#\?[[:space:]]*\(dbbutton\.visible.*\)/dbbutton.visible = True/g" /etc/openerp/$branch/web-client/$name.conf
-fi
-
-log_echo "Creating openerp-web log files..."
-touch /var/log/openerp/$name/web-client-access.log
-touch /var/log/openerp/$name/web-client-error.log
+#Change log permissions
 chown -R openerp_$name:openerp /var/log/openerp/$name
 chmod 664 /var/log/openerp/$name/*.log
 
@@ -285,9 +293,11 @@ log_echo "Creating apache rewrite file..."
 cp -a /etc/openerp/apache2/ssl-$branch-skeleton /etc/openerp/apache2/rewrites/$name
 sed -i "s#\\[NAME\\]#$name#g" /etc/openerp/apache2/rewrites/$name
 
-if [[ $branch == "6.1" ]] || [[ $branch == "trunk" ]]; then
+if [[ "$branch" =~ ^6\.1|trunk$ ]]; then
+    #Use OpenERP embedded XML-RPC Werzeug server
     sed -i "s#\\[PORT\\]#20$port#g" /etc/openerp/apache2/rewrites/$name >> $INSTALL_LOG_FILE
 else
+    #Use OpenERP Web client
     sed -i "s#\\[PORT\\]#23$port#g" /etc/openerp/apache2/rewrites/$name >> $INSTALL_LOG_FILE
 fi
 
@@ -298,23 +308,24 @@ mkdir -p /var/run/openerp/$name
 
 #~ Start server now
 if [[ $start_now =~ ^[Yy]$ ]]; then
-	log_echo "Starting openerp server and web client..."
-	service postgresql$postgresql_init start
-	service apache2 restart
-	service openerp-server-$name start
-	service openerp-web-$name start
-	log_echo ""
+    log_echo "Starting openerp server and web client..."
+    service postgresql$postgresql_init start
+    service apache2 restart
+    service openerp-server-$name start
+    #Start OpenERP Web client if not embedded
+    if [[ "$branch" =~ ^5\.0|6\.0$ ]]; then service openerp-web-$name start; fi
+    log_echo ""
 fi
 
 #~ Make developer menus
 #if [[ $server_type == "station" ]]; then
-#	sudo -E -u $openerp_user ccorp-openerp-mkmenus $branch $name
+#    sudo -E -u $openerp_user ccorp-openerp-mkmenus $branch $name
 #fi
 
 #~ Add server to hosts file if station
 
 if [[ $server_type == "station" ]]; then
-	echo "127.0.1.1	$name.localhost" >> /etc/hosts
+    echo "127.0.1.1    $name.localhost" >> /etc/hosts
 fi
 
 install_change_perms
